Creating a react-powered drop-down
1. Create a new pack called pay_type that’ll be the root of our implementation.
  app/javascript/packs/pay_type.jsx

2. Create the PayTypeSelector component that we’ll use to replace the existing
pay type selector drop-down.
  /app/javascript/PayTypeSelector/index.jsx

3. Bring the component into our checkout view using javascript_pack_tag() and a piece of markup that React can hook into in order to render the component.
  app/views/orders/new.html.erb
  app/views/orders/_form.html.erb

The core concept in React is components. A component is a view, backed by some sort of state. When the state changes, the view rerenders. The view can behave differently depending on the current state inside the component. For us, we’ll track the currently selected pay type as our state and have our view render different input tags based on that.