Update PayTypeSelector (react code) to dynamically render correct form
A React component is a view and state; and when state changes, the view is rerendered by calling the component’s render() method. We want the view to be rerendered when the user changes payment types, so we need to get the currently selected payment type into the component’s state, which can be done via setState().