Iter K3: Translating Checkout
Getting i18n-js to work requires a bit of configuration

First, we’ll configure i18n-js to convert our translations. This is done by a middleware that the gem provides. We can set this up by adding a line of code to config/application.rb

Next, we need to tell Rails to serve up the translations that i18n-js provides. We also need to make the I18n object available. We can do that by adding two require directives to app/assets/javascripts/application.js.

The last bit of configuration we need for i18n-js is to tell it what the currently chosen locale is. We can do that by rendering a dynamic script tag in our application layout in app/views/layouts/application.html.erb.

With the set ups above, we can now add calls to I18n.t() inside the JSX of our React components.